requirements.txt
streamlit
pillow
python-dotenv
requests
beautifulsoup4
feedparser
pymupdf
google-generativeai